clean up and refactoring. no change in logic
