--- src/yodapa/plugin.py
@@ -1,66 +1,0 @@
-import importlib
-import pkgutil
-from pathlib import Path
-from typing import List, Tuple
-
-import typer
-
-
-def discover_plugins() -> List[Tuple[str, typer.Typer]]:
-    """Discover plugins in the 'plugins' directory and the local plugins directory."""
-
-    plugins: List[Tuple[str, typer.Typer]] = list()
-
-    # 1. Discover plugins within the 'plugins' directory
-    plugins_pkg = importlib.import_module('yodapa.plugins')
-    plugins_path = plugins_pkg.__path__
-
-    for finder, name, ispkg in pkgutil.iter_modules(plugins_path):
-        # print("finder", finder, "name", name, "ispkg", ispkg)
-        try:
-            module = importlib.import_module(f'yodapa.plugins.{name}')
-            if hasattr(module, "app") and isinstance(module.app, typer.Typer):
-                plugin_app: typer.Typer = module.app
-                plugins.append((name, plugin_app))
-        except Exception as e:
-            typer.echo(f"Failed to load plugin {name}: {e}", err=True)
-
-    # 2. Discover plugins in the local plugins directory
-    local_plugins_dir = Path.home() / ".yoda" / "plugins"
-    if local_plugins_dir.exists() and local_plugins_dir.is_dir():
-        for finder, name, ispkg in pkgutil.iter_modules([str(local_plugins_dir)]):
-            # print("finder", finder, "name", name, "ispkg", ispkg)
-            try:
-                module_path = local_plugins_dir / f"{name}.py"
-                if not module_path.exists():
-                    typer.echo(f"Plugin module {name} does not exist at {module_path}", err=True)
-                    continue
-
-                # Load the module from the file path
-                spec = importlib.util.spec_from_file_location(name, str(module_path))
-                if spec is None:
-                    typer.echo(f"Could not load spec for module {name}", err=True)
-                    continue
-
-                module = importlib.util.module_from_spec(spec)
-                spec.loader.exec_module(module)
-
-                if hasattr(module, "app") and isinstance(module.app, typer.Typer):
-                    plugin_app: typer.Typer = module.app
-                    plugins.append((name, plugin_app))
-            except Exception as e:
-                typer.echo(f"Failed to load local plugin {name}: {e}", err=True)
-
-    # uncomment to debug
-    # print("Plugins discovered:", plugins)
-    return plugins
-
-
-def load_plugins(app: typer.Typer, plugins: List[Tuple[str, typer.Typer]]):
-    """Load the plugins into the yoda typer app."""
-    for name, plugin in plugins:
-        try:
-            app.add_typer(plugin, name=name)
-            # typer.echo(f"Loaded plugin: {plugin.name}")
-        except Exception as e:
-            typer.echo(f"Error loading plugin {name}: {e}", err=True)
